+++ packages/linode-js-sdk/src/one-click-apps/types.ts
@@ -1,20 +1,24 @@
 export interface OneClickApp {
     deployments_active: number;
     id: number;
-    user_gravatar_id: string;
     label: string;
     description: string;
+    summary: string;
     images: string[];
     deployments_total: number;
     username: string;
     is_public: boolean;
     created: string;
     updated: string;
-    rev_note: string;
     script: string;
     user_defined_fields: UserDefinedField[];
     ordinal: number;
-    logo_url: string;
+    color_logo_url: string;
+    grey_logo_url: string;
+    related_guides: Doc[];
+    tips: string[];
+    related_info: Doc[];
+    href?: string;
 }
 
 export interface UserDefinedField {
@@ -25,3 +29,9 @@ export interface UserDefinedField {
     manyof?: string;
     default?: string;
 }
+
+export interface Doc {
+    title: string;
+    href: string;
+}
+
